+++ src/MeritooMenuBundle.php
@@ -10,6 +10,8 @@ declare(strict_types=1);
 
 namespace Meritoo\MenuBundle;
 
+use Meritoo\MenuBundle\DependencyInjection\Compiler\VisitorFactoryPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
@@ -21,4 +23,12 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  */
 class MeritooMenuBundle extends Bundle
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function build(ContainerBuilder $container): void
+    {
+        parent::build($container);
+        $container->addCompilerPass(new VisitorFactoryPass());
+    }
 }